+++ notendb/foot.php
@@ -8,7 +8,7 @@
     <a href="help.php?title=Hilfe" tabindex="-1">Hilfe</a>  |
     <a href="tests.php?title=Tests" tabindex="-1">Tests</a> |
     <a href="https://github.com/birrei/ndb/tree/main" tabindex="-1" target="_blank">GitHub</a>
-    <a href="https://github.com/birrei/ndb/tree/main" tabindex="-1" target="_blank">GitHub</a>
+    <a href="https://github.com/birrei/ndb/blob/main/changelog.md" tabindex="-1" target="_blank">Changelog</a>
 
 </p>  
 <hr>